Added automated subscription auction start
