+++ frame/rws/Cargo.toml
@@ -12,6 +12,7 @@ sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v
 sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.10", default-features = false }
 frame-system = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.10", default-features = false }
 frame-support = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.10", default-features = false }
+pallet-robonomics-staking = { path = "../staking", default-features = false } 
 
 [dev-dependencies]
 sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.10" }
@@ -28,4 +29,5 @@ std = [
     "sp-runtime/std",
     "frame-system/std",
     "frame-support/std",
+    "pallet-robonomics-staking/std",
 ]
