+++ frame/staking/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "pallet-robonomics-staking"
 description = "Stake your token to get some network bandwidth"
-version = "0.2.0"
+version = "0.3.0"
 authors = ["Airalab <research@aira.life>"]
 edition = "2018"
